requirements.txt
boto3
kubernetes
pyOpenSSL
python-json-logger
